[refactor] major simplification
--------------------------------------------------
Warnings: untested.
removes option to have unordered key sequences.
may be more processor intensive.
--------------------------------------------------
Basic Explanation:
Replaces the core logic with a ring array.
input events are added to the ring unconditionally,
and then if the escape condition is satisfied, triggers the
escape function in the pre-command-hook,
and can trigger the evil-escape-hook in the post-command-hook.

--------------------------------------------------
Outstanding issues:

--------------------------------------------------
Notes: